move 'registry/util/network' package into 'pkg/network'
